tests/wpt/web-platform-tests/service-workers/service-worker/ServiceWorkerGlobalScope/registration-attribute.https.html

<!DOCTYPE html>
<title>ServiceWorkerGlobalScope: registration</title>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src='../resources/test-helpers.sub.js'></script>
<script>

promise_test(function(t) {
    var script = 'resources/registration-attribute-worker.js';
    var scope = 'resources/scope/registration-attribute';

    return service_worker_unregister_and_register(t, script, scope)
      .then(function(registration) {
          return wait_for_state(t, registration.installing, 'activated');
        })
      .then(function() { return with_iframe(scope); })
      .then(function(frame) {
          var expected_events_seen = [
            'updatefound',
            'install',
            'statechange(installed)',
            'statechange(activating)',
            'activate',
            'statechange(activated)',
            'fetch',
          ];

          assert_equals(
              frame.contentDocument.body.textContent,
              expected_events_seen.toString(),
              'Service Worker should respond to fetch');
          frame.remove();
          return service_worker_unregister_and_done(t, scope);
        });
  }, 'Verify registration attribute on ServiceWorkerGlobalScope');

</script>
